Seeders / Panels / Heroes Crud
Criação dos Seedes
Template do Panel
Início do CRUD do Heroes

+++ database/migrations/2018_03_14_174221_create_heroes_table.php
@@ -21,7 +21,7 @@ class CreateHeroesTable extends Migration
             $table->integer('lifePoints');
             $table->integer('defensePoints');
             $table->integer('damagePoints');
-            $table->integer('attackSpeed');
+            $table->decimal('attackSpeed', 5, 2);
             $table->integer('movementSpeed');
             $table->integer('idUserInsert')->nullable();
             $table->integer('idUserUpdate')->nullable();
